@@ -1,18 +1,19 @@
-import Navbar from '../components/Navbar';
-import '../styles/app.css';
 import { UserProvider } from '@auth0/nextjs-auth0';
+import NavBar from '../components/NavBar';
+
+import '../styles/app.css';
 
 function MyApp({ Component, pageProps }) {
-    return (
-        <UserProvider>
-            <div className="bg-red-600 w-full p-10 min-h-screen">
-                <Navbar />
-                <div className="max-w-2xl mx-auto">
-                    <Component {...pageProps} />
-                </div>
-            </div>
-        </UserProvider>
-    );
+  return (
+    <UserProvider>
+      <div className="bg-blue-600 w-full p-10 min-h-screen">
+        <div className="max-w-2xl mx-auto">
+          <NavBar />
+          <Component {...pageProps} />
+        </div>
+      </div>
+    </UserProvider>
+  );
 }
 
 export default MyApp;
