@@ -1,24 +1,24 @@
 import { createSnippet } from '../../utils/Fauna';
+
 import { withApiAuthRequired, getSession } from '@auth0/nextjs-auth0';
 
 export default withApiAuthRequired(async function handler(req, res) {
-    const session = getSession(req, res);
-    const userId = session.user.sub;
-    const { code, language, description, name } = req.body;
-    if (req.method !== 'POST') {
-        return res.status(405).json({ msg: 'Method not allowed' });
-    }
-    try {
-        const createdSnippet = await createSnippet(
-            code,
-            language,
-            description,
-            name,
-            userId
-        );
-        return res.status(200).json(createdSnippet);
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ msg: 'Something went wrong.' });
-    }
+  const { sub: userID } = getSession(req, res).user;
+  const { code, language, description, name } = req.body;
+
+  if (req.method !== 'POST') {
+    return res.status(405).json({ msg: 'Method not allowed' });
+  }
+  try {
+    const snippet = await createSnippet(
+      code,
+      language,
+      description,
+      name,
+      userID
+    );
+    return res.status(200).json(snippet);
+  } catch (err) {
+    return res.status(500).json({ msg: 'Something went wrong.' });
+  }
 });
