@@ -1,26 +1,19 @@
 import { deleteSnippet } from '../../utils/Fauna';
+
 import { withApiAuthRequired, getSession } from '@auth0/nextjs-auth0';
 
 export default withApiAuthRequired(async function handler(req, res) {
-    const session = getSession(req, res);
-    const userId = session.user.sub;
-    
-    if (req.method !== 'DELETE') {
-        return res.status(405).json({ msg: 'Method not allowed' });
-    }
+  // TODO: Need to deal with userID; only correct user can delete their snippet
+  // const { sub: userID } = getSession(req, res).user;
+  const { id } = req.body;
+  if (req.method !== 'DELETE') {
+    return res.status(405).json({ msg: 'Method not allowed' });
+  }
 
-   
-    const { id } = req.body;
-    const existingRecord = await getSnippetById(id);
-    if (!existingRecord || existingRecord.data.userId !== userId) {
-        res.statusCode = 404;
-        return res.json({ msg: 'Record not found' });
-    }
-    try {
-        const deleted = await deleteSnippet(id);
-        return res.status(200).json(deleted);
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ msg: 'Something went wrong.' });
-    }
-}
+  try {
+    const snippet = await deleteSnippet(id);
+    return res.status(200).json(snippet);
+  } catch (err) {
+    return res.status(500).json({ msg: 'Something went wrong.' });
+  }
+});
