@@ -5,8 +5,8 @@ module.exports = {
   },
   purge: [],
   theme: {
-    extend: {},
+    extend: {}
   },
   variants: {},
-  plugins: [],
-}
+  plugins: []
+};
